pyproject.toml
[project]
name = "llm-agent-smith"
version = "0.1.0"
description = ""
authors = [
    {name = "danilo",email = "natan.danilo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0" # <-- Esta é a linha corrigida!
dependencies = [
    # Suas dependências serão adicionadas aqui pelo Poetry
    "langchain-google-genai (>=2.1.8,<3.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langgraph (>=0.5.3,<0.6.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)"
]

[tool.poetry]
packages = [{include = "llm_agent_smith", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"